07_Family_Readiness_Deployment_Support/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
